Modify make targets to run lint tests from Docker
This commit updates all the make targets for all the lint tests so that
the tests are run from a Docker container instead of the local system.
These tests require an unreleased version of the CFT lint image that
include the make helper functions and the wrapper scripts that call
them.